Fix Java JNI warnings (#4229)

+++ java/src/jni/h5eImp.c
@@ -455,7 +455,7 @@ Java_hdf_hdf5lib_H5_H5Eget_1msg(JNIEnv *env, jclass clss, jlong msg_id, jintArra
         H5_LIBRARY_ERROR(ENVONLY);
     namePtr[buf_size] = '\0';
 
-    theArray[0] = error_msg_type;
+    theArray[0] = (jint)error_msg_type;
 
     if (NULL == (str = ENVPTR->NewStringUTF(ENVONLY, namePtr)))
         CHECK_JNI_EXCEPTION(ENVONLY, JNI_FALSE);
